Fix serial read race in writeBlock()
With Serial.read(), any delay in recieving bytes would result in -1
being written to flash, and, when the bytes finally arrive, they would
cause data offset and communication errors.

This was fine with the TeensySPIFlasher.py client due to the way it
buffers CMD_SPI_WRITE_BLOCK and the initial data. But, the timing is
fragile, relying on all data bytes being ready when needed.